+++ x/posts/internal/keeper/keeper.go
@@ -175,13 +175,15 @@ func (k Keeper) getAnswersStoreKey(postID types.PostID, answerer sdk.AccAddress)
 	return []byte(types.PollAnswersStorePrefix + postID.String() + answerer.String())
 }
 
+// Save the post poll answers made by the answerer into the kv store
 func (k Keeper) SavePollPostAnswers(ctx sdk.Context, postID types.PostID, answers []uint64, answerer sdk.AccAddress) {
 	store := ctx.KVStore(k.StoreKey)
 
 	store.Set(k.getAnswersStoreKey(postID, answerer), k.Cdc.MustMarshalBinaryBare(&answers))
 }
 
-func (k Keeper) GetPollPostUserAnswers(ctx sdk.Context, postID types.PostID, user sdk.AccAddress) (pa []uint64) {
+// Returns all the answers that a user made to a poll post if they exists.
+func (k Keeper) GetPollPostUserAnswers(ctx sdk.Context, postID types.PostID, user sdk.AccAddress) (postAnswers []uint64) {
 	store := ctx.KVStore(k.StoreKey)
 
 	key := k.getAnswersStoreKey(postID, user)
@@ -190,9 +192,45 @@ func (k Keeper) GetPollPostUserAnswers(ctx sdk.Context, postID types.PostID, use
 		return []uint64{}
 	}
 
-	k.Cdc.MustUnmarshalBinaryBare(store.Get(key), &pa)
+	k.Cdc.MustUnmarshalBinaryBare(store.Get(key), &postAnswers)
 
-	return pa
+	return postAnswers
+}
+
+// Returns the total number of answers for the poll associated to the post with the given postID
+// It assumes that the postID is related to an existing post with poll data
+func (k Keeper) GetPollTotalAnswersAmount(ctx sdk.Context, postID types.PostID) sdk.Int {
+	store := ctx.KVStore(k.StoreKey)
+	iterator := sdk.KVStorePrefixIterator(store, k.getAnswersStoreKey(postID, nil))
+
+	totalVotes := 0
+	for ; iterator.Valid(); iterator.Next() {
+		var answers []uint64
+		k.Cdc.MustUnmarshalBinaryBare(iterator.Value(), &answers)
+		totalVotes += len(answers)
+	}
+
+	return sdk.NewInt(int64(totalVotes))
+}
+
+// Returns the total number of votes that an answer, associated to the answerID provided, gained in the poll associated to
+// the given postID.
+func (k Keeper) GetAnswerTotalVotes(ctx sdk.Context, postID types.PostID, answerID uint64) sdk.Int {
+	store := ctx.KVStore(k.StoreKey)
+	iterator := sdk.KVStorePrefixIterator(store, k.getAnswersStoreKey(postID, nil))
+
+	answerTotalVotes := 0
+	for ; iterator.Valid(); iterator.Next() {
+		var answers []uint64
+		k.Cdc.MustUnmarshalBinaryBare(iterator.Value(), &answers)
+		for _, answer := range answers {
+			if answer == answerID {
+				answerTotalVotes++
+			}
+		}
+	}
+
+	return sdk.NewInt(int64(answerTotalVotes))
 }
 
 // ClosePollPost retrieves a previously opened poll post and close it.