+++ x/posts/internal/types/keys.go
@@ -23,7 +23,10 @@ const (
 	ActionRemovePostReaction = "remove_post_reaction"
 
 	// Queries
-	QuerierRoute = ModuleName
-	QueryPost    = "post"
-	QueryPosts   = "posts"
+	QuerierRoute        = ModuleName
+	QueryPost           = "post"
+	QueryPosts          = "posts"
+	QueryPollUserAnswer = "poll-user-answers"
+	QueryAnswersAmount  = "poll-answers-amount"
+	QueryAnswerVotes    = "poll-answer-votes"
 )